fix: 냉장고 파먹기 레시피 조회 시 재료 일치도 계산 오류 수정

+++ src/main/java/com/recipe/app/src/recipe/application/RecipeService.java
@@ -259,7 +259,7 @@ public class RecipeService {
                                     || ingredientNamesInFridge.contains(ingredient.getIngredientName()))
                             .count();
 
-                    return recipeIngredientInFridgeCnt / recipeIngredientCnt * 100;
+                    return (long) ((double) recipeIngredientInFridgeCnt / recipeIngredientCnt * 100);
                 }));
     }
